scraper/requirements.txt
selenium
pdfplumber
beautifulsoup4
requests
python-dotenv
supabase-py 